Added branding and favicon
Added a branding and favicon element to the build. Favicon’s are
popular on desktop based websites, but now there are many options for
the mobile devices on the market. I found a great site that will take
your image and convert into all mobile formats as well as favicon, and,
it gives you the corresponding code to go with it! Check it out here:
http://www.favicon-generator.org/